# perf: skip empty js hook for js plugin