<!-- Thank you for contributing! -->


## With 1000 empty plugin(only have name), use `10000` case, 
#### Main 
```bash
Benchmark 1: node --run build:rolldown
  Time (mean ± σ):     953.0 ms ±   6.0 ms    [User: 15170.3 ms, System: 875.5 ms]
  Range (min … max):   947.8 ms … 959.5 ms    3 runs
```

chrome-trace json size: 
![image](https://github.com/user-attachments/assets/459f8b51-beec-4d6e-991e-511a9fe859cd)
This is only partial trace json file, I need to terminate it because it stuck my computer.

#### This branch
```bash
Benchmark 1: node --run build:rolldown
  Time (mean ± σ):     584.0 ms ±   6.7 ms    [User: 2775.2 ms, System: 1057.3 ms]
  Range (min … max):   577.2 ms … 590.5 ms    3 runs
```
Tracing chrome-json: 161MB


## With zero plugin
#### Main
```bash
Benchmark 1: node --run build:rolldown
  Time (mean ± σ):     548.0 ms ±   6.9 ms    [User: 2439.8 ms, System: 926.7 ms]
  Range (min … max):   542.4 ms … 555.7 ms    3 runs
```

#### This branch
```bash
Benchmark 1: node --run build:rolldown
  Time (mean ± σ):     537.3 ms ±  12.8 ms    [User: 2450.7 ms, System: 1034.6 ms]
  Range (min … max):   527.7 ms … 551.9 ms    3 runs
```




<!-- Please insert your description here and provide especially info about the "what" this PR is solving -->